refactor: move membership and character fetch into new service
