+++ src/app/services/bungie-api.service.ts
@@ -33,12 +33,10 @@ import {
   DestinyCollectiblesComponent,
   DestinyItemInvestmentStatDefinition,
 } from "bungie-api-ts/destiny2";
-import { getMembershipDataForCurrentUser } from "bungie-api-ts/user";
 import { AuthService } from "./auth.service";
 import { HttpClient } from "@angular/common/http";
 import { DatabaseService } from "./database.service";
 import { environment } from "../../environments/environment";
-import { BungieMembershipType } from "bungie-api-ts/common";
 import { IManifestArmor } from "../data/types/IManifestArmor";
 import {
   IInventoryArmor,
@@ -49,6 +47,7 @@ import { ArmorSlot } from "../data/enum/armor-slot";
 import { ArmorPerkOrSlot } from "../data/enum/armor-stat";
 import { ConfigurationService } from "./configuration.service";
 import { IManifestCollectible } from "../data/types/IManifestCollectible";
+import { MembershipService } from "./membership.service";
 
 // This will mutate the incoming armor item to add the relevant stats
 // plugHashes is received as a simple array of hashes,
@@ -102,7 +101,8 @@ export class BungieApiService {
     private authService: AuthService,
     private http: HttpClient,
     private db: DatabaseService,
-    private config: ConfigurationService
+    private config: ConfigurationService,
+    private membership: MembershipService
   ) {
     this.config.configuration.subscribe(async (config) => {
       this.config_assumeEveryLegendaryIsArtifice = config.assumeEveryLegendaryIsArtifice;
@@ -160,37 +160,12 @@ export class BungieApiService {
       });
   }
 
-  async getCharacters() {
-    let destinyMembership = await this.getMembershipDataForCurrentUser();
-    if (!destinyMembership) {
-      await this.authService.logout();
-      return [];
-    }
-
-    const profile = await getProfile((d) => this.$http(d), {
-      components: [DestinyComponentType.Characters],
-      membershipType: destinyMembership.membershipType,
-      destinyMembershipId: destinyMembership.membershipId,
-    });
-
-    return (
-      Object.values(profile?.Response.characters.data || {}).map((d) => {
-        return {
-          characterId: d.characterId,
-          clazz: d.classType as DestinyClass,
-          emblemUrl: d.emblemBackgroundPath,
-          lastPlayed: Date.parse(d.dateLastPlayed),
-        };
-      }) || []
-    );
-  }
-
   async transferItem(
     itemInstanceId: string,
     targetCharacter: string,
     equip = false
   ): Promise<boolean> {
-    let destinyMembership = await this.getMembershipDataForCurrentUser();
+    let destinyMembership = await this.membership.getMembershipDataForCurrentUser();
     if (!destinyMembership) {
       await this.authService.logout();
       return false;
@@ -244,7 +219,7 @@ export class BungieApiService {
 
   async moveItemToVault(itemInstanceId: string) {
     console.info("moveItemToVault", itemInstanceId);
-    let destinyMembership = await this.getMembershipDataForCurrentUser();
+    let destinyMembership = await this.membership.getMembershipDataForCurrentUser();
     if (!destinyMembership) {
       await this.authService.logout();
       return;
@@ -267,58 +242,6 @@ export class BungieApiService {
     };
 
     await transferItem((d) => this.$httpPost(d), payload);
-  }
-
-  async getMembershipDataForCurrentUser() {
-    var membershipData = JSON.parse(localStorage.getItem("auth-membershipInfo") || "null");
-    var membershipDataAge = JSON.parse(localStorage.getItem("auth-membershipInfo-date") || "0");
-    if (membershipData && Date.now() - membershipDataAge < 1000 * 60 * 60 * 24) {
-      console.log("getMembershipDataForCurrentUser -> loading cached! ");
-      return membershipData;
-    }
-
-    console.info("BungieApiService", "getMembershipDataForCurrentUser");
-    let response = await getMembershipDataForCurrentUser((d) => this.$http(d));
-    let memberships = response?.Response.destinyMemberships;
-    console.info("Memberships:", memberships);
-    memberships = memberships.filter(
-      (m) => m.crossSaveOverride == 0 || m.crossSaveOverride == m.membershipType
-    );
-    console.info("Filtered Memberships:", memberships);
-
-    let result = null;
-    if (memberships?.length == 1) {
-      // This guardian only has one account linked, so we can proceed as normal
-      result = memberships?.[0];
-    } else {
-      // This guardian has multiple accounts linked.
-      // Fetch the last login time for each account, and use the one that was most recently used.
-      let lastLoggedInProfileIndex: any = 0;
-      let lastPlayed = 0;
-      for (let id in memberships) {
-        const membership = memberships?.[id];
-        const profile = await getProfile((d) => this.$http(d), {
-          components: [DestinyComponentType.Profiles],
-          membershipType: membership.membershipType,
-          destinyMembershipId: membership.membershipId,
-        });
-        if (!!profile && profile.Response?.profile.data?.dateLastPlayed) {
-          let date = Date.parse(profile.Response?.profile.data?.dateLastPlayed);
-          if (date > lastPlayed) {
-            lastPlayed = date;
-            lastLoggedInProfileIndex = id;
-          }
-        }
-      }
-      console.info(
-        "getMembershipDataForCurrentUser",
-        "Selected membership data for the last logged in membership."
-      );
-      result = memberships?.[lastLoggedInProfileIndex];
-    }
-    localStorage.setItem("auth-membershipInfo", JSON.stringify(result));
-    localStorage.setItem("auth-membershipInfo-date", JSON.stringify(Date.now()));
-    return result;
   }
 
   // Collect the list of unlocked exotic armor pieces
@@ -357,7 +280,7 @@ export class BungieApiService {
           (1000 * 3600) / 2
         )
           return;
-    let destinyMembership = await this.getMembershipDataForCurrentUser();
+    let destinyMembership = await this.membership.getMembershipDataForCurrentUser();
     if (!destinyMembership) {
       await this.authService.logout();
       return;
